a lot of changes to pp. two more 1d pp indices were added. the major change is in tour2d pp. the two existing indices were modified to be calculated on raw data rather than sphered data. they will still work with sphered data. pp will now start up without having sphered data.

+++ types.h
@@ -180,7 +180,7 @@ typedef struct {
 
   /* temporary space */
   gint *nright, *index;
-  gfloat *x;
+  gdouble *x;
 
 } cartgini_param;
 
@@ -189,7 +189,7 @@ typedef struct {
 
   /* temporary space */
   gint *nright, *index;
-  gfloat *x;
+  gdouble *x;
 
 } cartentropy_param;
